Refactored Tests and added some more.
